backend/requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
defusedxml==0.7.1
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
djoser==2.2.0
flake8==7.3.0
gunicorn==21.2.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
Pillow==10.0.1
psycopg2==2.9.9
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
